add base class for course list fragment
