+++ app/src/main/java/org/stepic/droid/base/FragmentActivityBase.java
@@ -2,6 +2,8 @@ package org.stepic.droid.base;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.support.annotation.IdRes;
+import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
@@ -85,5 +87,11 @@ public abstract class FragmentActivityBase extends AppCompatActivity {
     protected void onDestroy() {
         super.onDestroy();
         ButterKnife.unbind(this);
+    }
+
+    protected void setFragment(@IdRes int res, Fragment fragment) {
+        android.support.v4.app.FragmentTransaction fragmentTransaction = getSupportFragmentManager().beginTransaction();
+        fragmentTransaction.replace(res, fragment);
+        fragmentTransaction.commit();
     }
 }
