+++ app/src/main/java/org/stepic/droid/view/activities/SearchCourseResultActivity.java
@@ -3,29 +3,20 @@ package org.stepic.droid.view.activities;
 import android.app.SearchManager;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.view.View;
-import android.widget.ListView;
+import android.support.v4.app.Fragment;
+import android.support.v7.widget.Toolbar;
 
 import org.stepic.droid.R;
 import org.stepic.droid.base.FragmentActivityBase;
+import org.stepic.droid.view.fragments.FindCoursesFragment;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
 
 public class SearchCourseResultActivity extends FragmentActivityBase {
 
-    @Bind(R.id.swipe_refresh_layout_mycourses)
-    protected SwipeRefreshLayout mSwipeRefreshLayout;
-
-    @Bind(R.id.list_of_courses)
-    protected ListView mListOfCourses;
-
-    @Bind(R.id.report_problem)
-    protected View mReportConnectionProblem;
-
-    @Bind(R.id.empty_courses)
-    protected View mEmptyCoursesView;
+    @Bind(R.id.toolbar)
+    Toolbar mToolbar;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -33,19 +24,30 @@ public class SearchCourseResultActivity extends FragmentActivityBase {
         setContentView(R.layout.fragment_courses);
         ButterKnife.bind(this);
 
+        initActivity();
+
         // Get the intent, verify the action and get the query
         Intent intent = getIntent();
         if (Intent.ACTION_SEARCH.equals(intent.getAction())) {
             String query = intent.getStringExtra(SearchManager.QUERY);
-            doMySearch(query);
+            doSearch(query);
         }
     }
 
-    private void doMySearch(String query) {
-        //just the stub
-        mListOfCourses.setVisibility(View.GONE);
-        mSwipeRefreshLayout.setVisibility(View.GONE);
-        mEmptyCoursesView.setVisibility(View.VISIBLE);
-        mReportConnectionProblem.setVisibility(View.GONE);
+    private void doSearch(String query) {
+
+    }
+
+    private void initActivity() {
+        Fragment fragment = new FindCoursesFragment();
+        setFragment(R.id.frame, fragment);
+
+        bus.register(this);
+    }
+
+    @Override
+    protected void onDestroy() {
+        bus.unregister(this);
+        super.onDestroy();
     }
 }
