+++ app/src/main/java/org/stepic/droid/view/fragments/CourseSearchFragment.java
@@ -1,0 +1,5 @@
+package org.stepic.droid.view.fragments;
+
+public class CourseSearchFragment extends CourseListFragmentBase {
+
+}
